feat: restructure files and create util folder
