+++ components/SummaryStats/SummaryStats.module.scss
@@ -1,4 +1,4 @@
-.summary {
+.summaryStats {
   border-collapse: collapse;
   width: 100%;
   caption {
